move template rendering related tests
