+++ packages/ember/tests/routing/model_loading_test.js
@@ -2,11 +2,7 @@
 import { Route } from '@ember/-internals/routing';
 import Controller from '@ember/controller';
 import { Object as EmberObject, A as emberA } from '@ember/-internals/runtime';
-import {
-  moduleFor,
-  ApplicationTestCase,
-  getTextOf,
-} from 'internal-test-helpers';
+import { moduleFor, ApplicationTestCase, getTextOf } from 'internal-test-helpers';
 import { run } from '@ember/runloop';
 import { computed, set } from '@ember/-internals/metal';
